Update CodingameVsMonacoInShadowDomComparison app, Add new SideBySideDiffEditor app
