+++ src/CodingameVsMonacoInShadowDomComparison.css
@@ -1,0 +1,24 @@
+h1, .editor-title {
+  padding: 0 4rem;
+}
+
+.editor-title {
+  font-size: 1rem;
+  height: 25px;
+}
+
+.h2-and-editor-container {
+  border: 1px solid #ccc;
+  margin: 1rem;
+  height: 200px;
+}
+
+.shadow-container {
+  height: 200px;
+}
+
+.editor-container {
+  overflow: hidden;
+  width: 100%;
+  height: 175px;
+}
